fix: rename getColonyPayouts to match function name
